+++ plugins/kyverno/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 0.1.2
+* Fix log message when checking for clusterreports
+* Include policy/clusterpolicy title and description in report
+* bash refactoring
+
 ## 0.1.1
 * add missing CRD check to Kyverno script, fix entrypoint for Dockerfile